pyproject.toml
[project]
name = "exemple-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.3.0",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "pyqt5>=5.15.11",
    "pyside2>=5.15.2.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]
